Break up importers
